# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Sarathi API
  version: 0.1.0
paths:
...
  /api/v1/tutor/messages:
    post:
      tags:
      - tutor
      summary: Send Tutor Message
      description: 21 POST /api/v1/tutor/messages - SSE stream
      operationId: send_tutor_message_api_v1_tutor_messages_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorMessageRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/tutor/threads/{id}:
    get:
      tags:
      - tutor
      summary: Get Tutor Thread
      description: 22 GET /api/v1/tutor/threads/{id}
      operationId: get_tutor_thread_api_v1_tutor_threads__id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TutorMessageRequest:
      properties:
        lesson_id:
          type: string
          format: uuid
          title: Lesson Id
        thread_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Thread Id
        content:
          type: string
          title: Content
        context_block_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Context Block Id
      type: object
      required:
      - lesson_id
      - content
      title: TutorMessageRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer